Conflicts:
	library/class_upfront_output.php
	styles/editor-interface.css
	styles/editor-interface.css.map

+++ library/servers/class_upfront_ajax.php
@@ -52,13 +52,13 @@ class Upfront_Ajax extends Upfront_Server {
 		$load_dev = $_POST['load_dev'] == 1 ? true : false;
 		$post_type = isset($_POST['new_post']) ? $_POST['new_post'] : false;
 		$parsed = false;
-		
+
 		//Check if assigned WP template and delete DB layout
 		if(isset($_POST['post_id']) && !empty($_POST['post_id']) && isset($_POST['data']['specificity']) && !empty($_POST['data']['specificity'])) {
 			$template = get_post_meta((int)$_POST['post_id'], '_wp_page_template', true);
 			$theme = Upfront_ChildTheme::get_instance();
 			$prefix = $theme->get_prefix();
-			if(!empty($template)) {
+			if(!empty($template) && $template != "default") {
 				delete_option($prefix.'-'.$_POST['data']['specificity']);
 			}
 		}
